requirements.txt
# Required packages and versions
psutil==6.0.0
pebble==5.0.7

# Storage
pandas==2.2.2
tables==3.9.2

# Math
scipy==1.14.0
mpmath==1.3.0
numpy==1.23.5

# Display
matplotlib~=3.9.1
seaborn==0.13.2

# Learning
tensorflow-hub==0.16.1
tensorflow_datasets==4.9.6
torch==2.4.1
absl-py~=2.1.0
scikit-learn~=1.5.1

# lakde
ninja==1.11.*

# Smoothing
scikit-misc==0.5.1
statsmodels==0.14.2

# Developement and Testing
pytest==8.3.2
pycodestyle==2.12.1
autopep8==2.3.1
